server/requirements.txt
Flask
Flask-Cors
pymongo
dnspython
python-dotenv
werkzeug
pyjwt
tensorflow
pillow
numpy
